namespace change to singular
