+++ src/Dto/Filter/LanguageFilter.php
@@ -2,27 +2,23 @@
 
 declare(strict_types=1);
 
-namespace OneSignal\Dto\Filters;
+namespace OneSignal\Dto\Filter;
 
-final class AppVersionFilter extends AbstractFilter
+final class LanguageFilter extends AbstractFilter
 {
-    public const GT = '>';
-
-    public const LT = '<';
-
     public const EQ = '=';
 
     public const NEQ = '!=';
 
     /**
-     * @var self::GT|self::LT|self::EQ|self::NEQ
+     * @var self::EQ|self::NEQ
      */
     protected string $relation;
 
     protected string $value;
 
     /**
-     * @param self::GT|self::LT|self::EQ|self::NEQ $relation
+     * @param self::EQ|self::NEQ $relation
      */
     public function __construct(string $relation, string $value)
     {
@@ -33,7 +29,7 @@ final class AppVersionFilter extends AbstractFilter
     public function toArray(): array
     {
         return [
-            'field' => 'app_version',
+            'field' => 'language',
             'relation' => $this->relation,
             'value' => $this->value,
         ];
