+++ src/Dto/Filter/SessionCountFilter.php
@@ -2,25 +2,29 @@
 
 declare(strict_types=1);
 
-namespace OneSignal\Dto\Filters;
+namespace OneSignal\Dto\Filter;
 
-final class LanguageFilter extends AbstractFilter
+final class SessionCountFilter extends AbstractFilter
 {
+    public const GT = '>';
+
+    public const LT = '<';
+
     public const EQ = '=';
 
     public const NEQ = '!=';
 
     /**
-     * @var self::EQ|self::NEQ
+     * @var self::GT|self::LT|self::EQ|self::NEQ
      */
     protected string $relation;
 
-    protected string $value;
+    protected int $value;
 
     /**
-     * @param self::EQ|self::NEQ $relation
+     * @param self::GT|self::LT|self::EQ|self::NEQ $relation
      */
-    public function __construct(string $relation, string $value)
+    public function __construct(string $relation, int $value)
     {
         $this->relation = $relation;
         $this->value = $value;
@@ -29,7 +33,7 @@ final class LanguageFilter extends AbstractFilter
     public function toArray(): array
     {
         return [
-            'field' => 'language',
+            'field' => 'session_count',
             'relation' => $this->relation,
             'value' => $this->value,
         ];
