+++ src/Dto/Segment/CreateSegment.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace OneSignal\Dto\Segments;
+namespace OneSignal\Dto\Segment;
 
 use OneSignal\Dto\AbstractDto;
-use OneSignal\Dto\Filters\AbstractFilter;
+use OneSignal\Dto\Filter\AbstractFilter;
 
 class CreateSegment implements AbstractDto
 {
